Extract C# symbol name normalizer
